requirements.txt
asgiref==3.5.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==4.1.4
django-allauth==0.50.0
django-cors-headers==3.13.0
django-extensions==3.2.1
django-filter==22.1
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
idna==3.4
Markdown==3.4.1
MarkupSafe==2.1.1
oauthlib==3.2.2
Pillow==9.3.0
psycopg2==2.9.5
pycparser==2.21
PyGreSQL==5.2.4
PyJWT==2.6.0
pyOpenSSL==22.1.0
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.13
Werkzeug==2.2.2
